update sleep generator

+++ src/main/java/fr/xebia/java8/refactoring/step5/repository/FakeRepositoryTools.java
@@ -1,9 +1,9 @@
 package fr.xebia.java8.refactoring.step5.repository;
 
 public class FakeRepositoryTools {
-    public static void waitInSecond(int seconds) {
+    public static void waitInSecond(double seconds) {
         try {
-            Thread.sleep(seconds * 1000);
+            Thread.sleep((long) seconds * 1000);
         } catch (InterruptedException e) {
             throw new RuntimeException(e);
         }
